voice-agent-day27/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
aiofiles
